requirements.txt
accelerate 
bitsandbytes
datasets==2.13.1
transformers == 4.40.0
peft == 0.10.0
trl
scipy
huggingface_hub == 0.22.2
python-dotenv == 0.13.0